requirements.txt
pdfkit
azure.storage.blob
azure.core
azure.search.documents
msal